sim2real_toolkit/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyarrow>=6.0.0
opencv-python>=4.5.0
av>=9.0.0
pillow>=9.0.0
scipy>=1.7.0
pyyaml>=5.4.0
tqdm>=4.62.0
matplotlib>=3.4.0
PySide6>=6.2.0
scikit-image>=0.19.0
albumentations>=1.3.0

